set_led_col function optimized and changed data type

+++ main/ADC.c
@@ -9,7 +9,7 @@ void adc_read(void *pvParameters) {
 
     while (1) {  
         // Read ADC value
-        int32_t adc_value = adc1_get_raw(ADC_CHANNEL);
+        int16_t adc_value = adc1_get_raw(ADC_CHANNEL);
 
         AdcValue = adc_value;  // Update the global adc_delay variable with the latest ADC reading
         vTaskDelay(pdMS_TO_TICKS(50));  // Delay for 50 milliseconds before the next reading
